feat: refactored the domain list
